fix(middleware): Added exception handling to throttle middleware

+++ src/traffik/backends/inmemory.py
@@ -16,7 +16,7 @@ class InMemoryBackend(
         typing.Optional[
             typing.MutableMapping[
                 str,
-                typing.MutableMapping[str, int],
+                typing.MutableMapping[str, float],
             ]
         ],
         HTTPConnectionT,
